requirements.txt
pandas>=2.1.0
numpy>=1.24.0
plotly>=5.18.0
streamlit>=1.28.0
sqlalchemy>=2.0.0
openai>=1.6.0
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0
scipy>=1.11.0
requests>=2.31.0
pytest>=7.4.0
